requirements.txt
sh
pyyaml
colorlog
